Added some simple preference generation and fixed (?) the csv path logging. Updated utility values and student decisions to something that makes more sense. Added console logs to every single agent interaction to see what is happening. The timetabler is mysteriously only receiving a single null message matching the mt for registration and then not receiving any at all even though the inform shows up in the sniffer. Truly i do not know what to do anymore

+++ src/Generation/TimetableGeneration.java
@@ -23,7 +23,7 @@ public class TimetableGeneration
                 int i = r.nextInt(emptyTimeslotIds.size()) ;
                 var timeslotId = (int) emptyTimeslotIds.get(i);
                 
-                tutorial.setTimeSlotId((long) timeslotId);
+                tutorial.setTimeSlotId(timeslotId);
                 
                 timetable.setTutorial(timeslotId, tutorial);
                 
@@ -32,4 +32,5 @@ public class TimetableGeneration
         });
         return timetable;
     }
+    
 }
